+++ NEWS.md
@@ -6,6 +6,7 @@
   information.
 - Added support for record element constraints and record fields with
   unconstrained array types.
+- Alias of multidimensional array allowed in VHDL-2008 mode.
 
 ## Version 1.6.2 - 2022-04-03
 - Fix `make -j` with GNU make (#440).